#24 - group 이름 변경으로 인한 import 변경
이전 프로젝트에서 사용했던 그룹 이름을 변경함으로써
파일들의 수정이 생김
