+++ src/main/java/com/plus120/meditime/controller/MediBoxController.java
@@ -1,0 +1,47 @@
+package com.plus120.meditime.controller;
+
+import java.util.List;
+
+import com.plus120.meditime.dto.MediBoxDTO;
+import com.plus120.meditime.dto.MediForUserDTO;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.plus120.meditime.service.MediBoxService;
+
+@CrossOrigin(origins = "*")
+@RestController
+@RequestMapping("/medibox")
+public class MediBoxController {
+    
+    private final MediBoxService mediBoxService;
+
+    public MediBoxController(MediBoxService mediBoxService) {
+        this.mediBoxService = mediBoxService;
+    }
+
+    @PostMapping()
+    public void insertMediBox(@RequestBody MediBoxDTO mediBoxDTO) {
+        System.out.println("이름" + mediBoxDTO.getItemName());
+        System.out.println("유저아이디" + mediBoxDTO.getUserId());
+        System.out.println("아이디" + mediBoxDTO.getId());
+        mediBoxService.insertMediBox(mediBoxDTO);
+    }
+
+    @GetMapping("/select")
+    public List<MediForUserDTO> selectMediBox(@RequestParam("userId") String UserId) {
+        System.out.println("컨트롤러" + UserId);
+        return mediBoxService.selectMediBox(UserId);
+    }
+
+    // @DeleteMapping()
+    // public void deleteMediBox(@RequestParam("userId") String userId) {
+    //     mediBoxService.deleteMediBox(userId);
+    // }
+
+}
