+++ src/main/java/com/plus120/meditime/controller/MemberController.java
@@ -1,0 +1,20 @@
+package com.plus120.meditime.controller;
+
+
+import com.plus120.meditime.service.MemberService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@CrossOrigin(origins = "*")
+@RequiredArgsConstructor
+@RestController
+@RequestMapping("/api")
+public class MemberController {
+
+    @Autowired
+    private MemberService memberService;
+
+}
