+++ src/main/java/com/plus120/meditime/domain/MedicineBox.java
@@ -1,0 +1,20 @@
+package com.plus120.meditime.domain;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+@Entity
+@Table(name="medicinebox")
+public class MedicineBox {
+    
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name="id")
+    private Long id;
+    
+    
+}
