+++ src/main/java/com/plus120/meditime/dto/LoginDTO.java
@@ -1,0 +1,10 @@
+package com.plus120.meditime.dto;
+
+import lombok.*;
+
+@Getter @Setter
+@NoArgsConstructor @AllArgsConstructor @Builder
+public class LoginDTO {
+    private String email;
+    private String password;
+}
