+++ src/main/java/com/plus120/meditime/dto/MemberInfo.java
@@ -1,0 +1,10 @@
+package com.plus120.meditime.dto;
+
+import lombok.*;
+
+@Getter @Setter
+@NoArgsConstructor @AllArgsConstructor @Builder
+public class MemberInfo {
+    private String email;
+    private String username;
+}
