+++ src/main/java/com/plus120/meditime/service/MedicineService.java
@@ -1,0 +1,7 @@
+package com.plus120.meditime.service;
+
+import com.plus120.meditime.dto.MedicineDTO;
+
+public interface MedicineService {
+    MedicineDTO getMedicine(String itemName, String efcItem);
+}
