+++ src/main/java/com/plus120/meditime/service/MemberService.java
@@ -1,0 +1,4 @@
+package com.plus120.meditime.service;
+
+public interface MemberService {
+}
