+++ src/main/java/com/plus120/meditime/service/MemberServiceImpl.java
@@ -1,0 +1,12 @@
+package com.plus120.meditime.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+@Transactional
+public class MemberServiceImpl implements MemberService {
+}
+
