--- src/main/java/dev/medi/meditime/controller/MemberController.java
@@ -1,20 +1,0 @@
-package dev.medi.meditime.controller;
-
-
-import dev.medi.meditime.service.MemberService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@CrossOrigin(origins = "*")
-@RequiredArgsConstructor
-@RestController
-@RequestMapping("/api")
-public class MemberController {
-
-    @Autowired
-    private MemberService memberService;
-
-}
