--- src/main/java/dev/medi/meditime/domain/Article.java
@@ -1,15 +1,0 @@
-package dev.medi.meditime.domain;
-
-import java.time.LocalDateTime;
-
-public class Article {
-    private Long id;
-    private String title; // 제목
-    private String content; // 본문
-    private String hashtag; // 해시태그
-
-    private LocalDateTime createdAt; // 생성일시
-    private String createdBy; // 생성자
-    private LocalDateTime modifiedAt; // 수정일시
-    private String modifiedBy; // 수정자
-}
