--- src/main/java/dev/medi/meditime/domain/ArticleComment.java
@@ -1,14 +1,0 @@
-package dev.medi.meditime.domain;
-
-import java.time.LocalDateTime;
-
-public class ArticleComment {
-    private Long id;
-    private Article article; // 게시글 (ID)
-    private String content;
-
-    private LocalDateTime createdAt; // 생성일시
-    private String createdBy; // 생성자
-    private LocalDateTime modifiedAt; // 수정일시
-    private String modifiedBy; // 수정자
-}
