--- src/main/java/dev/medi/meditime/dto/JoinDTO.java
@@ -1,11 +1,0 @@
-package dev.medi.meditime.dto;
-
-import lombok.*;
-
-@Getter @Setter
-@NoArgsConstructor @AllArgsConstructor @Builder
-public class JoinDTO {
-    private String email;
-    private String password;
-    private String nickname;
-}
