--- src/main/java/dev/medi/meditime/dto/LoginDTO.java
@@ -1,10 +1,0 @@
-package dev.medi.meditime.dto;
-
-import lombok.*;
-
-@Getter @Setter
-@NoArgsConstructor @AllArgsConstructor @Builder
-public class LoginDTO {
-    private String email;
-    private String password;
-}
