--- src/main/java/dev/medi/meditime/dto/MemberInfo.java
@@ -1,10 +1,0 @@
-package dev.medi.meditime.dto;
-
-import lombok.*;
-
-@Getter @Setter
-@NoArgsConstructor @AllArgsConstructor @Builder
-public class MemberInfo {
-    private String email;
-    private String username;
-}
