--- src/main/java/dev/medi/meditime/repository/MedicineRepository.java
@@ -1,17 +1,0 @@
-package dev.medi.meditime.repository;
-
-import java.util.List;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-
-import dev.medi.meditime.domain.Medicine;
-
-public interface MedicineRepository extends JpaRepository<Medicine, String>{
-
-    @Query("select m from Medicine m where m.itemName like %:itemName% or m.efcItem like %:efcItem%")
-    List<Medicine> searchMediList(@Param("itemName") String itemName, @Param("efcItem") String efcItem );
-    
-}
-
