--- src/main/java/dev/medi/meditime/repository/MemberRepository.java
@@ -1,18 +1,0 @@
-package dev.medi.meditime.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-
-import dev.medi.meditime.domain.Member;
-
-import java.util.Optional;
-
-public interface MemberRepository extends JpaRepository<Member, Long>{
-
-    Optional<Member> findByEmail(String email);
-
-    Optional<Member> findByUsername(String username);
-
-    @Query("select m from Member m join fetch m.authorities a where m.username = :username")
-    Optional<Member> findByUsernameWithAuthority(String username);
-}
