--- src/main/java/dev/medi/meditime/service/MediBoxService.java
@@ -1,16 +1,0 @@
-package dev.medi.meditime.service;
-
-import java.util.List;
-
-import dev.medi.meditime.dto.MediBoxDTO;
-import dev.medi.meditime.dto.MediForUserDTO;
-
-public interface MediBoxService {
-    
-    void insertMediBox(MediBoxDTO mediBoxDTO);
-
-    // void deleteMediBox(String userId);
-
-    List<MediForUserDTO> selectMediBox(String userId);
-
-}
