--- src/main/java/dev/medi/meditime/service/MedicineService.java
@@ -1,7 +1,0 @@
-package dev.medi.meditime.service;
-
-import dev.medi.meditime.dto.MedicineDTO;
-
-public interface MedicineService {
-    MedicineDTO getMedicine(String itemName, String efcItem);
-}
