--- src/main/java/dev/medi/meditime/service/MemberService.java
@@ -1,4 +1,0 @@
-package dev.medi.meditime.service;
-
-public interface MemberService {
-}
