--- src/main/java/dev/medi/meditime/service/MemberServiceImpl.java
@@ -1,12 +1,0 @@
-package dev.medi.meditime.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-@Transactional
-public class MemberServiceImpl implements MemberService {
-}
-
